Trying to figure out why parts of the quad tree are disappearing.
I've gone through a number of tests, and I can't tell when nodes will be
in one frame but not the next.  There are over a hundred nodes when
things are busy, and I can't debug my through thousands of particles and
over a hundred nodes to see when they subdivide.  I've already tried to
track when nodes are active on one frame but not on the next, but I
couldn't get it to work.  Hmm...