Figured out why some of the nodes seemed to be disappearing.
After a lot of digging around, I found that the number of active
particles according to the quad tree was not matching the number of
active particles being reported by the atomic counter in
ParticleUpdate.comp.  Finally I found the problem:

At the beginning of ParticleQuadTree::AddParticlestoTree(...), I copy
the particle array into a local particle array for better caching as the
tree is being populated.

I did this:
memcpy(_localParticleArray, particleCollection, numParticles);

Whoops.  Corrected to this:
memcpy(_localParticleArray, particleCollection, sizeof(Particle) *
numParticles);