Success!  Experiment with calculating quadrant indices with float-to-int shenanigans completed.
Thanks to a co-worker for the algorithm.  I had an issue when I
misunderstood the result of "is top" as "is bottom".  After a fair
amount of debugging, and listening to a lot of Critical Role episode 82
(Deadly Echos), I got it working.  The next goal is to expand this
algorithm to the rest of node selection and to gather all the particles
from neighboring nodes into a single array.