Well that didn't work.
I tried to have the quad tree class do extra duty by generating an array
of particle collision candidates after it generated the tree.  It
followed the same algorithm that the ParticleCollisions compute shader
is using to populate the collidableParticleArray, but it bombed in
performance compared to the compute shader version.  So I'll be going
back to that one.