Moar cleanup!
I don't think that there is anything that I can do to improve
performance now.  My algorithm requires that particles be updated before
they are put into a node.  The GPU is terrible at that, so data needs to
be downloaded to the CPU.  Then the CPU makes the quad tree and uploads
it.  But that isn't my barrier.  My barrier is the particle collision
resolution in ParticleCollisions.comp.  The collision detection doesn't
seem to sap much performance in comparison.  I don't know why, but
that's all I can do.  On to particle-solid collisions!